non-admin fix hopefully

+++ app/controllers/admin/content_controller.rb
@@ -34,24 +34,28 @@ class Admin::ContentController < Admin::BaseController
   end
 
   def merge
-    curr_id = params["id"]
-    source_id = params["merge_with"]
+  
+    if @current_user.admin? do
+      curr_id = params["id"]
+      source_id = params["merge_with"]
 # debugger
     
-    curr_article = Article.find_by_id(curr_id)
-    source_article = Article.find_by_id(source_id)
+      curr_article = Article.find_by_id(curr_id)
+      source_article = Article.find_by_id(source_id)
     
-    if source_article != nil
-      curr_article.body = curr_article.body + "\r\n" + source_article.body
+      if source_article != nil
+        curr_article.body = curr_article.body + "\r\n" + source_article.body
       
-      # current_article.comments << source_article.comments
+        # current_article.comments << source_article.comments
 # debugger
       
-      curr_article.save!
-      source_article.destroy
+        curr_article.save!
+        source_article.destroy
+      end
     end
 # debugger
     redirect_to :action => 'index'
+    
   end
 
   def edit
